% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{add_fit}
\alias{add_fit}
\title{Add observed and fit values to plotly visualization}
\usage{
add_fit(
  p,
  x = NULL,
  yline = NULL,
  ymarker = NULL,
  ymin = NULL,
  ymax = NULL,
  ...,
  data = NULL,
  inherit = TRUE
)
}
\description{
Add observed and fit values to plotly visualization
}
