% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{dmuniform}
\alias{dmuniform}
\title{Mollified uniform density using a Gaussian distribution}
\usage{
dmuniform(x, lower = 0, upper = 1, sd = 0.1, log = FALSE)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{lower}{Lower range for the distribution.}

\item{upper}{Upper range for the distribution.}

\item{sd}{Standard deviation to control the shape of the distribution. A value of 0 will
be equivalent to a uniform distribution and values approach the normal as sd
approach infinity.}
}
\description{
Density distribution function for a mollified uniform using the
             Gaussian distribution.
}
\examples{
x <- seq(-1, 1, length = 100)
d <- dmuniform(x, lower = -0.5, upper = 0.5, sd = 0.05)
plot(x, d, type = "l", ylab = "Density")

}
