% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{logit}
\alias{logit}
\alias{inv_logit}
\title{Logit and inverse logit functions}
\usage{
logit(x, shift = FALSE)

inv_logit(x, shift = FALSE)
}
\arguments{
\item{x}{Probability or logit to transform}

\item{shift}{Shift transformation such that the domain is between -1 and 1? Defaults to domain of 0 to 1.}
}
\description{
Logit and inverse logit functions
}
\examples{

logit(0.9)
inv_logit(2.197225)

}
