% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{multispic}
\alias{multispic}
\title{Fit a multispecies surplus production model}
\usage{
multispic(
  inputs,
  center = TRUE,
  log_K_option = par_option(),
  log_B0_option = par_option(),
  log_r_option = par_option(),
  log_sd_B_option = par_option(),
  log_q_option = par_option(),
  log_sd_I_option = par_option(),
  logit_rho_option = par_option(),
  logit_phi_option = par_option(),
  species_cor = "none",
  temporal_cor = "none",
  K_groups = NULL,
  pe_covariates = NULL,
  n_forecast = 0,
  leave_out = NULL,
  start_par = NULL,
  light = FALSE,
  silent = FALSE
)
}
\arguments{
\item{inputs}{List that includes the following data.frames with required columns in
parentheses: \code{landings} (\code{species}, \code{year}, \code{landings}), \code{index} (\code{species}, \code{year},
\code{survey}, \code{index}). Covariates can be included in the \code{landings} data.frame
and specified using the \code{covariates} arguments (optional).}

\item{center}{Center input values to aid convergence?}

\item{log_K_option}{Settings for the estimation of \code{log_K}; define using \code{\link[=par_option]{par_option()}}.}

\item{log_B0_option}{Settings for the estimation of the starting biomass;
define using \code{\link[=par_option]{par_option()}}.}

\item{log_r_option}{Settings for the estimation of \code{log_r}; define using \code{\link[=par_option]{par_option()}}.}

\item{log_sd_B_option}{Settings for the estimation of sd for the process; define using
\code{\link[=par_option]{par_option()}}.}

\item{log_q_option}{Settings for the estimation of \code{log_q}; define using \code{\link[=par_option]{par_option()}}.}

\item{log_sd_I_option}{Settings for the estimation of sd for the indices; define using
\code{\link[=par_option]{par_option()}}.}

\item{logit_rho_option}{Setting for the estimation of the correlation across stocks; define using
\code{\link[=par_option]{par_option()}}.}

\item{logit_phi_option}{Setting for the estimation of temporal correlation in the process errors;
define using \code{\link[=par_option]{par_option()}}.}

\item{species_cor}{Correlation structure across species (\code{rho}). \code{"none"} will not estimate
correlations across species, \code{"one"} will estimate one shared correlation
parameter across species, and \code{"all"} will estimate correlation parameters
across all combinations of species.}

\item{temporal_cor}{Correlation structure across time (\code{phi}). \code{"none"} assumes no temporal dependence
in the process errors, \code{"rw"} assumes a random walk, and \code{"ar1"} fits an
AR1 process, estimating an extra parameter.}

\item{K_groups}{Formula pointing to the grouping variable to use to estimate \code{K} and
aggregate biomass in the production equation. Biomass from all species
(stocks) will be aggregated and one \code{K} value estimated if set to \code{NULL}.}

\item{pe_covariates}{Formula describing relationship between surplus production (process error)
and covariates. No covariates are applied if set to \code{NULL}.}

\item{n_forecast}{Number of years to forecast. Assumes status quo landings and covariates
(i.e. terminal values assumed through projected years).}

\item{leave_out}{Specific index values to leave out from the analysis (row number).
Useful for cross-validation. All data are kept if \code{NULL}.}

\item{start_par}{List of starting parameter values. Start parameters are internally
defined, however, it may be useful to supply parameters from a previous
model fit to speed up convergence.}

\item{light}{Skip running \code{\link[TMB:sdreport]{TMB::sdreport()}} and limit output to speed things up?}

\item{silent}{Disable tracing information?}
}
\description{
Fit a multispecies surplus production model
}
