% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_post}
\alias{plot_post}
\title{Plot parameter distributions}
\usage{
plot_post(
  prior_mean = 0,
  prior_sd = 1,
  show_prior = TRUE,
  post_mean = 0,
  post_sd = 1,
  names = NULL,
  length_out = 1000,
  xlab = "x",
  ylab = "Density",
  title = NULL,
  min_den = 1e-05,
  trans_fun = NULL
)
}
\arguments{
\item{prior_mean}{Mean of the prior.}

\item{prior_sd}{SD of the prior.}

\item{show_prior}{Display prior distribution?}

\item{post_mean}{Mean of the posterior.}

\item{post_sd}{SD of the posterior.}

\item{names}{Names of the posteriors (and priors) when multiple values are provided.}

\item{length_out}{Number of points to plot (more points will result in smoother curves).}

\item{xlab}{Label for the x axis.}

\item{ylab}{Label for the y axis.}

\item{title}{Plot title.}

\item{min_den}{Drop points with densities below this threshold.}

\item{trans_fun}{Function for transforming values (e.g. \verb{[exp]}, \verb{[inv_logit]}). Ignored if \code{NULL}.}
}
\description{
Plot posterior and prior (optional) distributions.
Currently limited to the normal distribution.
}
