% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_prior_post}
\alias{plot_prior_post}
\title{Plot prior and posterior distributions}
\usage{
plot_prior_post(x_min = -10, x_max = 10, prior_mean = 0,
  prior_sd = 1, post_mean = 0, post_sd = 1,
  post_names = "Posterior", length_out = 500, xlab = "x",
  ylab = "Density", title = NULL, min_den = 1e-04,
  exponentiate = FALSE)
}
\arguments{
\item{x_min}{Lower range of x values to plot}

\item{x_max}{Upper range of x values to plot}

\item{prior_mean}{Mean of the prior}

\item{prior_sd}{SD of the prior}

\item{post_mean}{Mean of the posterior (can be one or more values)}

\item{post_sd}{SD of the posterior (can be one or more values)}

\item{post_names}{Names for the posterior}

\item{length_out}{Number of points to plot (more points will result in smoother curves)}

\item{xlab}{Label for the x axis}

\item{ylab}{Label for the y axis}

\item{title}{Plot title}

\item{min_den}{Drop points with densities below this threshold}

\item{exponentiate}{Exponentiate?}
}
\description{
Plot prior and posterior distributions.
Currently limited to the normal distribution
}
\examples{


}
