% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{run_loo}
\alias{run_loo}
\title{Function for running leave one out cross-validation}
\usage{
run_loo(fit)
}
\arguments{
\item{fit}{Object from \code{\link[=fit_model]{fit_model()}}}
}
\value{
Returns a list of four:
\enumerate{
\item \code{fit}  -  all fit objects from each step,
\item \code{obs}  -  log observations that were left out at each step
\item \code{pred} -  log predictions at each step, and
\item \code{mse}  -  mean squared error of the predictions (leave one out cross validation score).
}
}
\description{
Function for running leave one out cross-validation
}
