% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{run_retro}
\alias{run_retro}
\title{Function for running a retrospective analysis}
\usage{
run_retro(fit, folds = 10, progress = TRUE)
}
\arguments{
\item{fit}{Object from \code{\link[=fit_model]{fit_model()}}}

\item{folds}{Number of years to 'fold' back}

\item{progress}{Display progress bar? (Generated using \link[progress:progress_bar]{progress::progress_bar})}
}
\value{
Returns a list with three objects:
\enumerate{
\item \code{retro_fits}  -  a list including a series of fits from each retrospective fold
\item \code{hindcasts} -  a data.frame with observed, but left out, and predicted indices from each fold
\item \code{mse}  -  mean squared error of the hindcasts (measure of forecasting skill)
}
}
\description{
Function for running a retrospective analysis
}
\details{
This function runs a retrospective analysis whereby terminal survey index estimates
are excluded from the analysis. A hindcast is also preformed in each fold where
observed values are left out but predicted to measure the models forecasting skill.
}
