% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{vis_multispic}
\alias{vis_multispic}
\title{Make a flexdashboard for visualizing the model fits}
\usage{
vis_multispic(fit, ...)
}
\arguments{
\item{fit}{Object produced by \code{\link[=multispic]{multispic()}}.}

\item{...}{Additional arguments to send to \code{\link[rmarkdown:run]{rmarkdown::run()}}}
}
\description{
Make a flexdashboard for visualizing the model fits
}
