% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_model}
\alias{fit_model}
\title{Fit a multispecies surplus production model}
\usage{
fit_model(inputs, survey_group = "gear_season",
  log_q_option = par_option(), cor_str = "one")
}
\arguments{
\item{inputs}{List that includes landings and index data}

\item{survey_group}{Name of column in the index data to group the survey parameter estimates by}

\item{log_q_option}{Settings defined using \code{\link{par_option}}.}

\item{cor_str}{Correlation structure across species. "none" will not estimate
correlations across species, "one" will estimate one shared correlation
parameter across species, and "all" will estimate correlation parameters
across all combinations of species.}
}
\description{
Fit a multispecies surplus production model
}
