% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_model}
\alias{fit_model}
\title{Fit a multispecies surplus production model}
\usage{
fit_model(inputs, survey_group = "gear_season",
  log_B0_option = par_option(), log_r_option = par_option(),
  log_sd_B_option = par_option(), log_q_option = par_option(),
  log_sd_I_option = par_option(), cor_str = "one", formula = NULL)
}
\arguments{
\item{inputs}{List that includes landings, index and covariate (optional) data.}

\item{survey_group}{Name of column in the index data to group the survey parameter estimates by}

\item{log_B0_option}{Settings for the estimation of the starting biomass;
define using \code{\link{par_option}}.}

\item{log_r_option}{Settings for the estimation of log_r; define using \code{\link{par_option}}.}

\item{log_sd_B_option}{Settings for the estimation of sd for the process; define using
\code{\link{par_option}}.}

\item{log_q_option}{Settings for the estimation of log_q; define using \code{\link{par_option}}.}

\item{log_sd_I_option}{Settings for the estimation of sd for the indices; define using
\code{\link{par_option}}.}

\item{cor_str}{Correlation structure across species. "none" will not estimate
correlations across species, "one" will estimate one shared correlation
parameter across species, and "all" will estimate correlation parameters
across all combinations of species.}

\item{formula}{Formula describing relationship between surplus production and covariates.
Not used if set to NULL.}
}
\description{
Fit a multispecies surplus production model
}
