% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_model}
\alias{fit_model}
\title{Fit a multispecies surplus production model}
\usage{
fit_model(
  inputs,
  scaler = sd(inputs$index$index),
  survey_group = "gear_season",
  log_K_option = par_option(),
  log_B0_option = par_option(),
  log_r_option = par_option(),
  log_sd_B_option = par_option(),
  log_q_option = par_option(),
  log_sd_I_option = par_option(),
  logit_rho_option = par_option(),
  logit_phi_option = par_option(),
  species_cor = "none",
  temporal_cor = "none",
  pe_formula = NULL,
  K_formula = NULL,
  leave_out = NULL,
  light = FALSE,
  silent = FALSE
)
}
\arguments{
\item{inputs}{List that includes landings, index and covariate (optional) data.}

\item{scaler}{Number to scale values by to aid convergence.}

\item{survey_group}{Name of column in the index data to group the survey parameter estimates by}

\item{log_K_option}{Settings for the estimation of log_K; define using \code{\link{par_option}}.}

\item{log_B0_option}{Settings for the estimation of the starting biomass;
define using \code{\link{par_option}}.}

\item{log_r_option}{Settings for the estimation of log_r; define using \code{\link{par_option}}.}

\item{log_sd_B_option}{Settings for the estimation of sd for the process; define using
\code{\link{par_option}}.}

\item{log_q_option}{Settings for the estimation of log_q; define using \code{\link{par_option}}.}

\item{log_sd_I_option}{Settings for the estimation of sd for the indices; define using
\code{\link{par_option}}.}

\item{logit_rho_option}{Setting for the estimation of the correlation across stocks; define using
\code{\link{par_option}}.}

\item{logit_phi_option}{Setting for the estimation of temporal correlation in the process errors;
define using \code{\link{par_option}}.}

\item{species_cor}{Correlation structure across species (rho). "none" will not estimate
correlations across species, "one" will estimate one shared correlation
parameter across species, and "all" will estimate correlation parameters
across all combinations of species.}

\item{temporal_cor}{Correlation structure across time (phi). "none" assumes no temporal dependence
in the process errors, "rw" assumes a random walk, and "ar1" fits an
AR1 process, estimating an extra parameter.}

\item{pe_formula}{Formula describing relationship between surplus production (process error)
and covariates. Not used if set to NULL.}

\item{K_formula}{Formula describing relationship between K and covariates. Not used if set
to NULL.}

\item{leave_out}{Specific index values to leave out from the analysis (row number).
Useful for cross-validation. All data are kept if NULL.}

\item{light}{Skip running sdreport and limit output to speed things up?}

\item{silent}{Disable tracing information?}
}
\description{
Fit a multispecies surplus production model
}
