% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{par_option}
\alias{par_option}
\title{Helper for defining the settings of a parameter}
\usage{
par_option(option = "fixed", mean = 0, sd = 1)
}
\arguments{
\item{option}{Should the parameter be estimated freely ("fixed") or revolve
around an estimated mean and sd ("random"), a fixed mean
("prior_mean") or a fixed mean and sd ("prior").}

\item{mean}{Mean value of the parameter. Treated as a starting value if
option is "random" or as a prior if option is "prior_mean" or
option is "prior". Ignored if option is "fixed".}

\item{sd}{SD value of the parameter. Treated as a starting value is
option is "random" or a prior if option is "prior". Ignored
if option is "fixed".}
}
\description{
Helper for defining the settings of a parameter
}
