% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{run_loo}
\alias{run_loo}
\title{Function for running leave one out cross-validation}
\usage{
run_loo(fit)
}
\arguments{
\item{fit}{Object from \code{\link{fit_model}}}
}
\value{
Returns a list of four: 1) fit - all fit objects from each setp,
        2) obs - log observations that were left out at each step
        3) pred - log predictions at each step, and 4) mse - mean squared
        error of the predictions (leave one out cross validation score).
}
\description{
Function for running leave one out cross-validation
}
