% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{run_loo}
\alias{run_loo}
\title{Function for running leave one out cross-validation}
\usage{
run_loo(fit, progress = TRUE)
}
\arguments{
\item{fit}{Object from \code{\link[=fit_model]{fit_model()}}}

\item{progress}{Display progress bar? (Generated using the progressr and progress packages)}
}
\value{
Returns a list with three objects:
\enumerate{
\item \code{obs}  -  log observations that were left out at each step
\item \code{pred} -  log predictions at each step, and
\item \code{mse}  -  mean squared error of the predictions (leave one out cross validation score).
}
}
\description{
Function for running leave one out cross-validation
}
